Fix gh#65: Narrow post lifecycle services
