+++ .clawpatch/features/feat_gh-65_4e1b171485.json
@@ -1,1 +1,93 @@
-{"schemaVersion":1,"featureId":"feat_gh-65_4e1b171485","title":"gh#65: R2.5 — Sweep: narrow PostServiceError + PostLifecycleAbort; remove dead inline logic","summary":"Imported from GitHub issue gh#65 in BACKLOG.md section \"Candidate 2 — Post Lifecycle aggregate\".","kind":"job","source":"github-issue-import","confidence":"high","entrypoints":[{"path":"packages/api/src/services/post.service.ts","symbol":null,"route":null,"command":null},{"path":"packages/worker/src/post-lifecycle.service.ts","symbol":null,"route":null,"command":null},{"path":"packages/shared/src/post","symbol":null,"route":null,"command":null}],"ownedFiles":[{"path":"packages/api/src/services/post.service.ts","reason":"mentioned by issue body"},{"path":"packages/worker/src/post-lifecycle.service.ts","reason":"mentioned by issue body"},{"path":"packages/shared/src/post","reason":"mentioned by issue body"},{"path":"packages/shared/src/post/aggregate.ts","reason":"mentioned by issue body"}],"contextFiles":[{"path":"docs/adr/0003-post-lifecycle-is-pure-functions-not-shared-repository.md","reason":"mentioned by issue body"},{"path":"docs/adr/0004-publish-cycle-orchestration-stays-in-the-worker.md","reason":"mentioned by issue body"},{"path":"BACKLOG.md","reason":"backlog queue item for gh#65"}],"tests":[],"tags":["github-issue","gh#65","github-url:https://github.com/joshgk00/social-media-scheduler/issues/65","backlog-order:0010","backlog-section:candidate-2-post-lifecycle-aggregate","label:enhancement"],"trustBoundaries":["network","filesystem","concurrency"],"status":"needs-fix","lock":null,"findingIds":["fnd_sig-gh-65-50d3a984eb_e157fbed1f"],"patchAttemptIds":[],"analysisHistory":[{"runId":"20260519T201935-c24f6a","kind":"github-issue-import","summary":"Imported GitHub issue gh#65","provider":"gh","model":null,"createdAt":"2026-05-19T20:19:35.980Z"}],"createdAt":"2026-05-19T20:19:35.980Z","updatedAt":"2026-05-19T20:19:35.980Z"}
+{
+  "schemaVersion": 1,
+  "featureId": "feat_gh-65_4e1b171485",
+  "title": "gh#65: R2.5 — Sweep: narrow PostServiceError + PostLifecycleAbort; remove dead inline logic",
+  "summary": "Imported from GitHub issue gh#65 in BACKLOG.md section \"Candidate 2 — Post Lifecycle aggregate\".",
+  "kind": "job",
+  "source": "github-issue-import",
+  "confidence": "high",
+  "entrypoints": [
+    {
+      "path": "packages/api/src/services/post.service.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    },
+    {
+      "path": "packages/worker/src/post-lifecycle.service.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    },
+    {
+      "path": "packages/shared/src/post",
+      "symbol": null,
+      "route": null,
+      "command": null
+    }
+  ],
+  "ownedFiles": [
+    {
+      "path": "packages/api/src/services/post.service.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/worker/src/post-lifecycle.service.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/shared/src/post",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/shared/src/post/aggregate.ts",
+      "reason": "mentioned by issue body"
+    }
+  ],
+  "contextFiles": [
+    {
+      "path": "docs/adr/0003-post-lifecycle-is-pure-functions-not-shared-repository.md",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "docs/adr/0004-publish-cycle-orchestration-stays-in-the-worker.md",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "BACKLOG.md",
+      "reason": "backlog queue item for gh#65"
+    }
+  ],
+  "tests": [],
+  "tags": [
+    "github-issue",
+    "gh#65",
+    "github-url:https://github.com/joshgk00/social-media-scheduler/issues/65",
+    "backlog-order:0010",
+    "backlog-section:candidate-2-post-lifecycle-aggregate",
+    "label:enhancement"
+  ],
+  "trustBoundaries": [
+    "network",
+    "filesystem",
+    "concurrency"
+  ],
+  "status": "fixed",
+  "lock": null,
+  "findingIds": [
+    "fnd_sig-gh-65-50d3a984eb_e157fbed1f"
+  ],
+  "patchAttemptIds": [],
+  "analysisHistory": [
+    {
+      "runId": "20260519T201935-c24f6a",
+      "kind": "github-issue-import",
+      "summary": "Imported GitHub issue gh#65",
+      "provider": "gh",
+      "model": null,
+      "createdAt": "2026-05-19T20:19:35.980Z"
+    }
+  ],
+  "createdAt": "2026-05-19T20:19:35.980Z",
+  "updatedAt": "2026-05-20T20:21:50.946Z"
+}
