+++ packages/worker/src/publish-worker.ts
@@ -203,31 +203,32 @@ export function createPublishHandler(deps: PublishHandlerDeps) {
       return { platformPostId: result.platformPostId };
     } catch (err) {
       if (err instanceof PostLifecycleAbort) {
+        const abortKind = err.invariant.kind;
         // Graceful aborts never throw to BullMQ — they resolve cleanly so
         // the job does not consume retry budget. The scanner's next pass
         // re-enqueues the post if it is still due.
-        if (err.reason === 'already_published') {
+        if (abortKind === 'already_published') {
           logger.info('Idempotent skip — post already published');
-          return { skipped: true, skipReason: err.reason };
+          return { skipped: true, skipReason: abortKind };
         }
         if (
-          err.reason === 'version_mismatch' ||
-          err.reason === 'budget_exhausted' ||
-          err.reason === 'not_scheduled' ||
-          err.reason === 'thread_unsupported' ||
-          err.reason === 'media_pending' ||
-          err.reason === 'token_unhealthy' ||
-          err.reason === 'rate_limit_exhausted'
+          abortKind === 'version_mismatch' ||
+          abortKind === 'budget_exhausted' ||
+          abortKind === 'not_scheduled' ||
+          abortKind === 'thread_unsupported' ||
+          abortKind === 'media_pending' ||
+          abortKind === 'token_unhealthy' ||
+          abortKind === 'rate_limit_exhausted'
         ) {
           // Phase 07-04 (TOKEN-05) / Phase 8 (LIMIT-06/LIMIT-07):
           // rate_limit_exhausted mirrors budget_exhausted / token_unhealthy —
           // post stays in `scheduled`, no retry budget consumed. The scanner
           // re-enqueues once the platform window resets.
           logger.info(
-            { reason: err.reason },
+            { reason: abortKind },
             'Graceful abort — scanner will re-evaluate',
           );
-          return { skipped: true, skipReason: err.reason };
+          return { skipped: true, skipReason: abortKind };
         }
       }
